% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zd_z0.R
\name{calculate_zd_z0}
\alias{calculate_zd_z0}
\title{calculate_zd_z0}
\usage{
calculate_zd_z0(
  lambda_building_pai,
  lambda_vegetaion_pai,
  P3D_summer = 0.2,
  P3D_winter = 0.6,
  P3D_intermediate = 0.4,
  lambda_building_fai,
  lambda_vegetaion_fai,
  Cdb = 1.2,
  lambda_combined_zHstd,
  lambda_combined_zH,
  lambda_combined_zHmax,
  timestamp,
  Kotthaus = F,
  wind_direction,
  wind_v = NULL
)
}
\arguments{
\item{lambda_building_pai}{building pai}

\item{lambda_vegetaion_pai}{vegetation pai}

\item{P3D_summer}{constant = 0.2}

\item{P3D_winter}{constant = 0.6}

\item{P3D_intermediate}{constant = 0.4}

\item{lambda_building_fai}{building fai}

\item{lambda_vegetaion_fai}{vegetation fai}

\item{Cdb}{constant = 1.2}

\item{lambda_combined_zHstd}{combined vegetation and building}

\item{lambda_combined_zH}{combined vegetation and building}

\item{lambda_combined_zHmax}{combined vegetation and building}

\item{timestamp}{timestamp}

\item{Kotthaus}{= F}

\item{wind_direction}{wind direction}

\item{wind_v}{= NULL}
}
\value{
the result is a data.frame with the calculater Zd and Zo.
}
\description{
This is a function calculate Zd and Zo using Kanda.
}
\examples{
# zo and zd calculation
ROTH_zd_z0 <- calculate_zd_z0(lambda_building_pai = Lambda_B_1deg$pai,
                            lambda_vegetaion_pai = Lambda_V_1deg$pai,
                            P3D_summer = 0.2,
                            P3D_winter = 0.6,
                            P3D_intermediate = 0.4,
                            lambda_building_fai = Lambda_B_1deg$fai,
                            lambda_vegetaion_fai = Lambda_V_1deg$fai,
                            Cdb = 1.2,
                            lambda_combined_zHstd = Lambda_C_1deg$zHstd,
                            lambda_combined_zH = Lambda_C_1deg$zH,
                            lambda_combined_zHmax = Lambda_C_1deg$zHmax,
                            timestamp = EC_ROTH$timestamp,
                            Kotthaus = F,
                            wind_direction = EC_ROTH$wd,
                            wind_v = NULL)

ROTH_zd_z0

}
