% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GIF.R
\name{create_GIF}
\alias{create_GIF}
\title{create_GIF}
\usage{
create_GIF(
  map,
  period = "month",
  gif_subtitle,
  water_mask,
  border,
  limits_scale = c(0, 100),
  breaks_scale = seq(0, 100, 20),
  colors_pallete,
  time_interval = 0.8,
  loops = 1,
  movie_name = "Urban_ET.gif"
)
}
\arguments{
\item{map}{sf map with with Urban ET per month derived from the map_urbanET()}

\item{period}{animated by "month" or "24hours", default = "month"}

\item{gif_subtitle}{map title, e.g. "Predicted ET (mm/month) - Berlin 2020"}

\item{water_mask}{sf object with water bodies to mask}

\item{border}{sf object with the city border}

\item{limits_scale}{scale limits of the map's legend, e.g. c(0,100)}

\item{breaks_scale}{map's legend breaks, e.g. seq(0, 100, 20)}

\item{colors_pallete}{continuous colour pallete for the map}

\item{time_interval}{time between each map in the animation default = .8 sec,}

\item{loops}{number of loops, default = 1}

\item{movie_name}{= name of the GIF file, default = "Urban_ET.gif"}
}
\value{
a GIF file saved in the R project directory.
}
\description{
This function animate the maps of Urban ET or cooling services according to a selected period.
}
\examples{
A animate Urban ET map monthly

Urban_ET_git <- create_GIF(map = Urban_ET_map_month[seq(3,25,2)],
                           period = "month",
                           gif_subtitle = "Predicted ET [mm/month] - Berlin 2020",
                           water_mask = water_polygons,
                           border = Berlin_border_utm,
                           limits_scale = c(0,100),
                           breaks_scale = seq(0, 100, 20),
                            colors_pallete = ETcolor,
                           time_interval = .8,
                           loops = 1,
                           movie_name = "Urban_ET.gif")

summary(Urban_ET_git)


}
