% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_parse_lateness.R
\name{find_and_parse_lateness}
\alias{find_and_parse_lateness}
\title{Find and Parse HH:MM:SS lateness columns in a Gradescope gradebook}
\usage{
find_and_parse_lateness(df, regex = "- Lateness", unit = "seconds")
}
\arguments{
\item{df}{The dataframe (gradebook)}

\item{regex}{The regular expression with which to match the titles of lateness columns}

\item{unit}{The unit to which to parse HH:MM:SS, in "seconds", "minutes", or "hours"}
}
\value{
Returns a dataframe with lateness columns parsed to specified unit
}
\description{
Find and Parse HH:MM:SS lateness columns in a Gradescope gradebook
}
\examples{
gradebook <- gradebook |> find_and_parse_lateness()
gradebook <- gradebook |> find_and_parse_lateness(unit = "minutes")
gradebook <- gradebook |> find_and_parse_lateness(unit = "hours")
gradebook <- gradebook |> find_and_parse_lateness(regex = "late_work", unit = "hours")
}
