% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fisbroker_map.R
\name{get_fisbroker_map}
\alias{get_fisbroker_map}
\title{get_fisbroker_map}
\usage{
get_fisbroker_map(
  url,
  crs_map = "+proj=utm +zone=33 +datum=WGS84 +units=m +no_defs"
)
}
\arguments{
\item{url}{map ulr from the Berlin Urban Atlas (fisbroker) site.}

\item{crs_map, }{projection default "+proj=utm +zone=33 +datum=WGS84 +units=m +no_defs".}
}
\value{
sf multipolygon map.
}
\description{
This is a function download maps from the Berlin Urban Atlas (fisbroker) site.
}
\examples{
split and save in different files to facilitate the processing
########
ETmap <- get_fisbroker_map("https://fbinter.stadt-berlin.de/fb/wfs/data/senstadt/s02_13whaus2017")

}
