% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LAI.R
\name{get_LAI_links}
\alias{get_LAI_links}
\title{get_LAI.R
This is a function download maps download the LAI 300m images from the Copernicus Vito portal.}
\usage{
get_LAI_links(
  user_VITO = NA,
  pass_VITO = NA,
  destfile = NA,
  start_date = NA,
  end_date = NA
)
}
\arguments{
\item{user_VITO}{user name from the Copernicus Vito portal}

\item{pass_VITO, }{password from the Copernicus Vito portal}

\item{destfile}{directory to download the file e.g. D:/data/lai/}

\item{start_date}{start date of the images , e.g. "2017-12-01"}

\item{end_date}{end date of the images}
}
\value{
list with links to download .nc file in a folder
}
\description{
get_LAI.R
This is a function download maps download the LAI 300m images from the Copernicus Vito portal.
}
\examples{
LAI_links <- get_LAI_links(user_VITO = 'your_user_name',
                           pass_VITO = '123456',
                           destfile = "D:/Data-Modelling/",
                           start_date = "2017-12-01",
                           end_date = "2022-09-20")

# download the images
for (i in 1:length(LAI_links)) {
               download.file(LAI_links[[1]][i],
               destfile = paste0("D:/Data-Modelling/LAI/",
                                 "LAI_", LAI_links[[2]][i],"_", i,".nc"),
               method="curl")
}

}
