% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name,
  soil_file = TRUE,
  name_simulation = TRUE,
  t = FALSE,
  Rin = FALSE,
  Rli = FALSE,
  p = FALSE,
  Ta = FALSE,
  u = FALSE,
  ea = FALSE,
  RH = FALSE,
  tts = FALSE,
  tto = FALSE,
  psi = FALSE,
  Cab = FALSE,
  Cca = FALSE,
  Cdm = FALSE,
  Cw = FALSE,
  Cs = FALSE,
  Cant = FALSE,
  SMC = FALSE,
  BSMBrightnes = FALSE,
  BSMlat = FALSE,
  BSMlon = FALSE,
  LAI = FALSE,
  hc = FALSE,
  LIDFa = FALSE,
  LIDFb = FALSE,
  z = FALSE,
  Ca = FALSE,
  Vcmax25 = FALSE,
  BallBerrySlope = FALSE,
  Cab_c = FALSE,
  Cca_c = FALSE,
  Cdm_c = FALSE,
  Cw_c = FALSE,
  Cs_c = FALSE,
  Cant_c = FALSE,
  cp = FALSE,
  Cbc = FALSE,
  Cp = FALSE,
  N = FALSE,
  rho_thermal = FALSE,
  tau_thermal = FALSE,
  Vcmax25_c = FALSE,
  BallBerrySlope_c = FALSE,
  BallBerry0 = FALSE,
  Type = FALSE,
  kV = FALSE,
  Rdparam = FALSE,
  Kn0 = FALSE,
  Knalpha = FALSE,
  Knbeta = FALSE,
  Tyear = FALSE,
  beta = FALSE,
  kNPQs = FALSE,
  qLs = FALSE,
  stressfactor = FALSE,
  Fluorescence = FALSE,
  fqe = FALSE,
  spectrum = FALSE,
  rss = FALSE,
  rs_thermal = FALSE,
  cs = FALSE,
  rhos = FALSE,
  lambdas = FALSE,
  BSMBrightness_c = FALSE,
  BSMlat_c = FALSE,
  BSMlon_c = FALSE,
  SMC_c = FALSE,
  LAI_c = FALSE,
  hc_c = FALSE,
  Rin_c = FALSE,
  Ta_c = FALSE,
  Rli_c = FALSE,
  p_c = FALSE,
  ea_c = FALSE,
  u_c = FALSE,
  tts_c = FALSE,
  tto_c = FALSE,
  psi_c = FALSE,
  z_c = FALSE,
  LIDFa_c = FALSE,
  LIDFb_c = FALSE,
  leafwidth = FALSE,
  Cv = FALSE,
  crowndiameter = FALSE,
  Ca_c = FALSE,
  Oa = FALSE,
  zo = FALSE,
  d = FALSE,
  Cd = FALSE,
  rb = FALSE,
  CR = FALSE,
  CD1 = FALSE,
  Psicor = FALSE,
  CSSOIL = FALSE,
  rbs = FALSE,
  rwc = FALSE,
  startDOY = TRUE,
  endDOY = TRUE,
  LAT = TRUE,
  LON = TRUE,
  timezn = TRUE,
  lite = TRUE,
  verify = FALSE,
  saveCSV = FALSE,
  mSCOPE = FALSE,
  simulation = FALSE,
  save_spectral = FALSE,
  calc_fluor = FALSE,
  calc_planck = FALSE,
  calc_xanthophyllabs = FALSE,
  Fluorescence_model = FALSE,
  calc_directional = FALSE,
  calc_vert_profiles = FALSE,
  soilspectrum = TRUE,
  applTcorr = TRUE,
  soil_heat_method = TRUE,
  calc_rss_rbs = TRUE,
  MoninObukhov = TRUE
)
}
\arguments{
\item{SCOPE_dir}{the diretory patch of SCOPE}

\item{Simulation_Name}{the name of the simulations}

\item{LAI, SMC, }{input variables, logical, if TRUE it will show the name of the input used for this parameter.}

\item{z_c, hc_c, }{input constant (_c), logical, if TRUE it will show the value of the input used for this parameter.}

\item{soil_heat_method}{model settings, logical, if TRUE it will show 0 for off and 1 or 2 for on.}
}
\value{
the result is a list with a table for each simulation with the input variables and seeting used.
}
\description{
This is a function to get all model inputs (variables and constants) and setting used in the simulations.
}
\examples{
Examples of uses of the get_parameters function
#######
SCOPE_parameters_pixel_1169 <- get_parameters(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name = "pixel_1169",
  LAI = TRUE,
  SMC = TRUE,
  hc = TRUE,
  hc_c = TRUE,
  z_c = TRUE)

SCOPE_parameters_pixel_1169[[1]]
length(SCOPE_parameters_pixel_1169 )

SCOPE_parameters_pixel_882 <- get_parameters(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name = "pixel_882",
  LAI = TRUE,
  SMC = TRUE,
  hc = TRUE,
  hc_c = TRUE,
  z_c = TRUE)

SCOPE_parameters_pixel_882
SCOPE_parameters_pixel_882[[1]]

length(SCOPE_parameters_pixel_882)

}
