% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\alias{get_accuracy}
\title{get_predictions}
\usage{
get_predictions(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name,
  output_file = "fluxes.csv",
  pred_vec = "lEtot"
)

get_accuracy(obs_vec, predictions, metric_function)
}
\arguments{
\item{SCOPE_dir}{the diretory patch of SCOPE}

\item{Simulation_Name}{simulation name stating the directory name generated by the run_SCOPE function.}

\item{output_file}{default "fluxes.csv" for the energy fluxes output}

\item{pred_vec}{desirble output variables.}

\item{obs_vec}{for get_accuracy a vector of observed values in the same timestamp as the predictions.}

\item{predictions}{the name of the file result of the get_predictions function.}

\item{metric_function}{for get_accuracy is possible to select the metric,}
}
\value{
The result is a data.frame with all the predicted output per simulation for get_predictions, and
        a table with the metrics per simulation for get_accuracy.
}
\description{
This is a function to get the predictions for a specific output simulated by SCOPE and calculate the accuracy.
}
\section{Functions}{
\itemize{
\item \code{get_accuracy}: function
}}

\examples{
Examples of uses of the get_predictions function
###############
library(tidyverse)

Predictions_pixel_1169 <- get_predictions(SCOPE_dir = "D:/SCOPE-master/",
                                          output_file = "fluxes.csv",
                                          pred_vec = "lEtot",
                                          Simulation_Name = "pixel_1169")

Predictions_pixel_1169

Examples of uses of the get_accuracy function

possible metrics: yardstick::rmse for Root mean squared error,
                  yardstick::mae for Mean absolute error,
                  yardstick::msd for Mean signed deviation,
                  yardstick::ccc for Concordance correlation coefficient,
                  yardstick::mase for Mean absolute scaled error - order by time
                  yardstick::rsq for R squared - correlation
                  yardstick::rsq_trad for R squared - traditional
                  yardstick::rpiq for Ratio of performance to inter-quartile
                  yardstick::rpd for Ratio of performance to deviation
                  yardstick::iic for Index of ideality of correlation
                  yardstick::mpe for Mean percentage error
                  yardstick::mape for Mean absolute percent error
                  yardstick::smape for Symmetric mean absolute percentage error

Get model accuracy for ET corrected
Predictions_metrics_1169 <- get_accuracy(obs_vec = EC_ROTH$ET_clean0,
                                         predictions = Predictions_pixel_1169,
                                         metric_function = yardstick::metric_set(yardstick::rsq,
                                                                                 yardstick::rmse,
                                                                                 yardstick::mae))
Predictions_metrics_1169

}
