% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reflectance.R
\name{get_reflectance}
\alias{get_reflectance}
\title{get_reflectance}
\usage{
get_reflectance(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name,
  output_file = "reflectance.csv"
)
}
\arguments{
\item{SCOPE_dir}{the diretory patch of SCOPE}

\item{Simulation_Name}{simulation name stating the directory name generated by the run_SCOPE function.}

\item{output_file}{default "reflectance.csv" output}
}
\value{
The result is a data.frame with all the predicted output per simulation name and date.
}
\description{
This is a function to get the reflectance for a specific output simulated by SCOPE.
}
\examples{
Examples of uses of the get_reflectance function
###############
library(tidyverse)

reflectance <- get_reflectance(SCOPE_dir = "D:/SCOPE-master/",
                                          Simulation_Name = "pixel_1169")

ggplot() +
geom_line(aes(y = reflectance[[1]]$X9, x = reflectance[[1]]$nm), colour="green") +
geom_line(aes(y = reflectance[[1]]$X35, x = reflectance[[1]]$nm), colour="darkgreen")

}
