% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SMCdata.R
\name{get_SMCdata}
\alias{get_SMCdata}
\title{get_SMCdata}
\usage{
get_SMCdata(
  lat_center,
  lon_center,
  radius_km,
  time_lag = "daily",
  meteo_var = "soil",
  start_date,
  end_date,
  data_dir = tempdir()
)
}
\arguments{
\item{lat_center}{latitude central point for the buffer}

\item{lon_center}{longitude central point for the buffer}

\item{radius_km}{for the buffer}

\item{time_lag}{default "hourly" , it can be change for daily or other timestamp available}

\item{meteo_var}{type of meteorological data target in the DWD ftp (see option below)}

\item{start_date}{select stations with date later than}

\item{end_date}{select stations with date earlier than}

\item{data_dir}{where to save the downlaod data, default temporary.}

\item{period}{default "historical", if change to "recent" you get one year lag up to "now".}

\item{var_name}{variable name target in the the downloaded meteo_var file (see option below)}
}
\value{
a list with: 1- a data.frame with the selected variable per station id for the set timestamp, and 2- a data.frame with metadata information, such as stations_id, start_date, end_date, station_height, latitude, longitude, stations_name, region, time_lag, variable, period, file, distance, url.
}
\description{
This is a function to download DWD data based on a buffer distance.
The meteorological variables are provide in data.frame per station id.
}
\examples{
#########
SMC_daily <- get_SMCdata(lat_center = 52.4537,
                         lon_center = 13.3017,
                         radius_km = 70,
                         time_lag = "daily",
                         meteo_var = "soil",
                         start_date = "2019-01-01",
                         end_date = "2020-12-31");

SMC_daily[[1]][[1]];
summary(SMC_daily[[1]][[1]]);

}
