% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Solardata.R
\name{get_Solardata}
\alias{get_Solardata}
\title{get_Solardata}
\usage{
get_Solardata(
  lat_center,
  lon_center,
  radius_km,
  time_lag = "hourly",
  meteo_var = "solar",
  start_date,
  end_date,
  data_dir = tempdir()
)
}
\arguments{
\item{lat_center}{latitude central point for the buffer}

\item{lon_center}{longitude central point for the buffer}

\item{radius_km}{for the buffer}

\item{time_lag}{default "hourly" , it can be change for daily or other timestamp available}

\item{meteo_var}{type of meteorological data target in the DWD ftp (see option below)}

\item{start_date}{select stations with date later than}

\item{end_date}{select stations with date earlier than}

\item{data_dir}{where to save the downlaod data, default temporary.}

\item{period}{default "historical", if change to "recent" you get one year lag up to "now".}

\item{var_name}{variable name target in the the downloaded meteo_var file (see option below)}
}
\value{
a list with: 1- a data.frame with the selected variable per station id for the set timestamp, and 2- a data.frame with metadata information, such as stations_id, start_date, end_date, station_height, latitude, longitude, stations_name, region, time_lag, variable, period, file, distance, url.
}
\description{
This is a function to download DWD data based on a buffer distance.
The meteorological variables are provide in data.frame per station id.
}
\examples{
#########
solar_radiation <- get_Solardata(lat_center = 52.4537,
                                 lon_center = 13.3017,
                                 radius_km = 70,
                                 time_lag = "hourly",
                                 meteo_var = "solar",
                                 start_date = "2018-12-31",
                                 end_date = "2021-01-01");

solar_radiation[[1]][[1]];
summary(solar_radiation[[1]][[1]]);

}
