% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_DWD.R
\name{interpolate_DWD}
\alias{interpolate_DWD}
\title{interpolate_DWD}
\usage{
interpolate_DWD(
  var_DWD,
  stations_DWD,
  res_utm,
  radius_km,
  crop_extent,
  vgm_model = c("Exp", "Mat", "Gau", "Sph")
)
}
\arguments{
\item{var_DWD}{a numeric vector with the DWD values per station and timestamp}

\item{stations_DWD}{the coordination of the DWD station}

\item{res_utm}{resolution of the final raster}

\item{radius_km}{the distance used in the get_dwd data to select the stations}

\item{crop_extent}{the extent of the area to crop}

\item{vgm_model}{the variagram theoretical model, default "Sph"}
}
\value{
stack raster with the climatological variable interpolated in space and time
}
\description{
This is a function to interpolate the DWD data downloaded.
The meteorological variables are point values and coordinates converted to a grid based on raster.
}
\examples{
# air temperature

krg_Ta <- lapply(1:n, FUN = function(i) Interp_DWD(var_DWD = as.numeric(Air_temp[[1]][i,-1]),
            stations_DWD = Air_temp[[2]],
            res = st_grid,
            crop_extent = raster::extent(Berlin_border_utm)));

krg_Ta[[1]];
plot(krg_Ta[[1]]);

}
