% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI_daily.R
\name{LAI_daily}
\alias{LAI_daily}
\title{LAI_daily}
\usage{
LAI_daily(star_data = NA, end_data = NA, LAI_rast = NA, crop_area = NA)
}
\arguments{
\item{LAI_rast}{raster with 10 day (+ gaps) LAI layers}

\item{crop_area, }{a sf polygon with desirable the extent}

\item{start_date}{start date of the images , e.g. "2017-12-01"}

\item{end_date}{end date of the images}
}
\value{
a raster with daily values of LAI
}
\description{
This is a function convert in daily values the downloaded LAI images for a location from the Copernicus Vito portal.
}
