% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inputs.R
\name{model_inputs}
\alias{model_inputs}
\title{model_inputs}
\usage{
model_inputs(SCOPE_dir)
}
\arguments{
\item{SCOPE_dir}{the directory of SCOPE.}
}
\value{
The result is a list with 3 object. the first is the input that vary in time or space.
The second is the constants and respectively default values. To change the constants in the run_scope functions
it is needed to include "_c" after the name, for instance, hc = 2 became hc_c = 10. The third are setting,
if set as zero (off) they are not applied when run SCOPE.
}
\description{
This is a function to get the model input parameters available (variables, constants values and settings).
}
\examples{
Examples of uses of the function
###########
model_inputs(SCOPE_dir = "D:/SCOPE-master/")

}
