% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_outputs.R
\name{model_outputs}
\alias{model_outputs}
\title{model_outputs}
\usage{
model_outputs(outputs = TRUE)
}
\arguments{
\item{outputs}{show the possible outputs and the settings to get them.}
}
\value{
The result is a table with the outputs and settings.
}
\description{
This is a function to get the model output available (variables and settings to get them).
}
\examples{
Examples of uses of the function

data(outputs_var)

outputs_var \%>\%
filter(simulation_file == "fluxes.csv")

}
