% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lateness.R
\name{parse_lateness}
\alias{parse_lateness}
\title{Parse \code{H*H:MM:SS} formatted lateness to seconds, minutes, or hours}
\usage{
parse_lateness(hms, unit = "seconds")
}
\arguments{
\item{hms}{A string of the form HH:MM:SS}

\item{unit}{The unit of time to be returned ("seconds", "minutes", or "hours")}
}
\value{
The converted time as a number
}
\description{
Parse \code{H*H:MM:SS} formatted lateness to seconds, minutes, or hours
}
\examples{
parse_lateness("1:00:00")
parse_lateness("1:00:00", unit = "seconds")
parse_lateness("12:00:00", unit = "minutes")
parse_lateness("0:10:00", unit = "hours")
}
