% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec_window.R
\name{prec_window}
\alias{prec_window}
\title{prec_window}
\usage{
prec_window(prec_vec, max_dry_hours, timestamp, value_on = 0.01)
}
\arguments{
\item{prec_vec}{a vector with the volume of precipitation.}

\item{max_dry_hours}{max possible number of hours without raining.}

\item{timestamp}{timestamp datetime.}
}
\value{
The result is a data.frame with all the number of hours after raining and the precipitaion volume in mm
}
\description{
This is a function to get the predictions for a specific output simulated by SCOPE and calculate the accuracy.
}
\examples{
Examples of uses of the function
library(tidyverse)

Input_Berlin_grid \%>\%
 filter(id_pixel == 1169) \%>\%
 mutate(timestamp=REddyProc::BerkeleyJulianDateToPOSIXct(t, tz = "UTC")) \%>\%
 select(timestamp, prec_hour) -> prec_1169

prec_window_1169 <- prec_window(prec_vec = prec_1169$prec_hour,
                              timestamp = prec_1169$timestamp,
                               max_dry_hours = 400)

}
