% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_assignments.R
\name{select_assignments}
\alias{select_assignments}
\title{Subset Gradebook by Assignment Titles}
\usage{
select_assignments(
  df,
  assignments,
  student_cols = c("Name", "SID", "Email", "Sections")
)
}
\arguments{
\item{df}{The data frame from which to select assignment columns}

\item{assignments}{Assignment titles to select}

\item{student_cols}{The list of columns that identify students in the gradebook}
}
\value{
A data frame of columns that match assignment titles, in order, alongside identifying student information
}
\description{
Subset Gradebook by Assignment Titles
}
\examples{
gradebook <- data.frame(
               `Name` = c("Carter Zenke", "Rongxin Liu"),
               `SID` = c("12345678", "87654321"),
               `Email` = c("test1@cs50.harvard.edu", "test2@cs50.harvard.edu"),
               `Sections` = c("Zenke", "Zenke"),
               `Problem Set 0` = c(9, 10),
               `Problem Set 0 - Max Points` = c(10, 10),
               `Problem Set 1` = c(2, 10),
               `Problem Set 1 - Max Points` = c(10, 10)
              )
pset0 <- gradebook |> select_assignments("Problem Set 0")
problems <- gradebook |> select_assignments(c("Problem Set 0", "Problem Set 1"))
}
