% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GIF.R
\name{create_GIF}
\alias{create_GIF}
\title{create_GIF}
\usage{
create_GIF(
  map,
  period = "month",
  gif_subtitle = "Predicted ET [mm/month] - Berlin 2020",
  water_mask = water_polygons,
  border = Berlin_border_utm,
  limits_scale = c(0, 100),
  breaks_scale = seq(0, 100, 20),
  colors_pallete = ETcolor,
  time_interval = 0.8,
  loops = 1,
  movie_name = "Urban_ET.gif"
)
}
\arguments{
\item{map}{a data.frame with SCOPE outputs, a datetime variable and pixel numbers (timestamp, id_pixel) from the get_prediction function}

\item{period}{= "month" to define the day to calculate the indices}

\item{gif_subtitle}{= "Predicted ET (mm/month) - Berlin 2020", variables to map, default c("ET", "Tsave"),}

\item{water_mask}{= water_polygons, the grip template (raster object, the same as the interpolation)}

\item{border}{= Berlin_border_utm, a vector the raster grid id_pixel masked and excluded from the SCOPE run}

\item{limits_scale}{= c(0,100), name of the sf polygon map object}

\item{breaks_scale}{= seq(0, 100, 20),  name of the vegetation fraction variable}

\item{colors_pallete}{= ETcolor, get the max ET form the 1km grid, default = 'max', faster and suitable in case of coarse grid}

\item{time_interval}{= .8,}

\item{loops}{= 1,}

\item{movie_name}{= "Urban_ET.gif"}
}
\value{
a GIF file saved in the R project directory.
}
\description{
This function animate the maps of Urban ET or cooling services according to a selected period.
}
\examples{
A animate Urban ET map monthly

Urban_ET_git <- create_GIF(map = Urban_ET_map_month[seq(3,25,2)],
                           period = "month",
                           gif_subtitle = "Predicted ET [mm/month] - Berlin 2020",
                           water_mask = water_polygons,
                           border = Berlin_border_utm,
                           limits_scale = c(0,100),
                           breaks_scale = seq(0, 100, 20),
                            colors_pallete = ETcolor,
                           time_interval = .8,
                           loops = 1,
                           movie_name = "Urban_ET.gif")

summary(Urban_ET_git)


}
