% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_and_parse_lateness.R
\name{find_and_parse_lateness}
\alias{find_and_parse_lateness}
\title{Find and Parse HH:MM:SS lateness columns in a Gradescope gradebook}
\usage{
find_and_parse_lateness(
  df,
  regex = "Lateness \\\\(H:M:S\\\\)",
  unit = "seconds"
)
}
\arguments{
\item{df}{The dataframe (gradebook)}

\item{regex}{The regular expression with which to match the titles of lateness columns}

\item{unit}{The unit to which to parse HH:MM:SS, in "seconds", "minutes", or "hours"}
}
\value{
Returns a dataframe with lateness columns parsed to specified unit
}
\description{
Find and Parse HH:MM:SS lateness columns in a Gradescope gradebook
}
\examples{
gradebook <- data.frame(
               `Problem Set 0: Scratch - Lateness (H:M:S)` = c("0:00:00", "1:32:11", "0:14:34")
             )
gradebook <- gradebook |> find_and_parse_lateness()
gradebook <- gradebook |> find_and_parse_lateness(unit = "minutes")
gradebook <- gradebook |> find_and_parse_lateness(unit = "hours")
gradebook <- gradebook |> find_and_parse_lateness(regex = "late_work", unit = "hours")
}
