% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{get_predictions}
\usage{
get_predictions(
  SCOPE_dir = "D:/SCOPE-master/",
  Simulation_Name,
  output_file = "fluxes.csv",
  pred_vec = "lEtot"
)
}
\arguments{
\item{SCOPE_dir}{the diretory patch of SCOPE}

\item{Simulation_Name}{simulation name stating the directory name generated by the run_SCOPE function.}

\item{output_file}{default "fluxes.csv" for the energy fluxes output}

\item{pred_vec}{desirble output variables.}
}
\value{
The result is a data.frame with all the predicted output per simulation name and date.
}
\description{
This is a function to get the predictions for a specific output simulated by SCOPE and calculate the accuracy.
}
\examples{
Examples of uses of the get_predictions function
###############
library(tidyverse)

Predictions_pixel_1169 <- get_predictions(SCOPE_dir = "D:/SCOPE-master/",
                                          output_file = "fluxes.csv",
                                          pred_vec = "lEtot",
                                          Simulation_Name = "pixel_1169")

Predictions_pixel_1169

}
